# Improve Etag ToString performance for DefaultFormat